Adjustments on simulator script + moving matter_qa to outside endpoint + handling  stress script location
